refactor: move payload calculations from PieChartView
