+++ app/src/main/java/otus/homework/customview/presentation/CustomViewViewModel.kt
@@ -10,6 +10,8 @@ import kotlinx.coroutines.launch
 import otus.homework.customview.R
 import otus.homework.customview.data.CustomViewRepository
 import otus.homework.customview.data.Payload
+import otus.homework.customview.presentation.model.PieChartAngle
+import otus.homework.customview.presentation.model.toPieChartAngles
 import javax.inject.Inject
 
 @HiltViewModel
@@ -20,9 +22,15 @@ class CustomViewViewModel @Inject constructor(
     private val _payloads = MutableStateFlow<List<Payload>>(emptyList())
     val payloads = _payloads.asStateFlow()
 
+    private val _pieChartAngles = MutableStateFlow<List<PieChartAngle>>(emptyList())
+    val pieChartAngles = _pieChartAngles.asStateFlow()
+
     init {
         viewModelScope.launch(dataLoadDispatcher) {
+
             _payloads.value = repository.loadPayloads(R.raw.payload)
+
+            _pieChartAngles.value = repository.loadPayloads(R.raw.payload).toPieChartAngles()
         }
     }
