+++ app/src/main/java/otus/homework/customview/presentation/model/PieChartAngle.kt
@@ -1,0 +1,15 @@
+package otus.homework.customview.presentation.model
+
+import otus.homework.customview.presentation.piechartview.PieChartColor
+
+data class PieChartAngle(
+    val id: Int,
+    val category: String,
+    val startAngle: Float,
+    val endAngle: Float,
+    val color: PieChartColor
+) {
+
+    val sweepAngle: Float
+        get() = endAngle - startAngle
+}
