+++ app/src/main/java/otus/homework/customview/presentation/piechartview/PieChartDrawer.kt
@@ -4,6 +4,7 @@ import android.graphics.Canvas
 import android.graphics.Paint
 import android.graphics.RectF
 import android.view.View
+import otus.homework.customview.presentation.model.PieChartAngle
 
 class PieChartDrawer(
     private val pieChartContainer: RectF,
