+++ app/src/main/java/otus/homework/customview/presentation/piechartview/PieChartView.kt
@@ -8,10 +8,12 @@ import android.graphics.RectF
 import android.os.Bundle
 import android.os.Parcelable
 import android.util.AttributeSet
+import android.util.Log
 import android.view.MotionEvent
 import android.view.View
-import otus.homework.customview.data.Payload
+import otus.homework.customview.presentation.model.PieChartAngle
 import otus.homework.customview.presentation.piechartview.GeometryHelper.isClickWithinPieChartAngle
+import otus.homework.customview.utils.TAG
 import otus.homework.customview.utils.dp
 import otus.homework.customview.utils.px
 import kotlin.math.min
@@ -40,20 +42,16 @@ class PieChartView @JvmOverloads constructor(
 
     private val pieChartContainer = RectF()
 
-    private var payloads: List<Payload> = emptyList()
-
     private val angles: MutableList<PieChartAngle> = mutableListOf()
 
     private var selectedSector: PieChartAngle? = null
-    private val angleCalculator = PieChartAngleCalculator()
     private val pieChartDrawer = PieChartDrawer(pieChartContainer, sectorPaint, selectionPaint)
 
     private var onCategorySelected : ((String) -> Unit)? = null
 
-    fun setPayloads(payloads: List<Payload>) {
-        this.payloads = payloads
-        angles.clear()
-        angles.addAll(angleCalculator.calculateAngles(payloads))
+    fun setPieChartAngles(angles : List<PieChartAngle>) {
+        this.angles.clear()
+        this.angles.addAll(angles)
         invalidate()
     }
